Issue #278 - removed commands that are no longer needed or used. Move countyapi.management.commands.utils.py to countyapi.utils.py and removed used functions.
