+++ countyapi/management/scraper/housing_location_info.py
@@ -1,8 +1,6 @@
-
 from django.db.utils import DatabaseError
 
-from countyapi.management.commands.utils import convert_to_int, join_with_space_and_convert_spaces, yesterday
-
+from countyapi.utils import convert_to_int, join_with_space_and_convert_spaces, yesterday
 from countyapi.models import HousingLocation
